feat(core): implement auto-refreshing of loaded entities
Previously when entity was loaded and we needed to reload it, providing
explicit `refresh: true` in the options was necessary. Refreshing of entity
also had one problematic side effect - the entity data (used for computing
change sets) was always updated based on the newly loaded entity, hence
forgetting the previous state (resulting in possibly lost updates done on
entity before refreshing).

Now we always merge the newly loaded data with current state, and when we
see updated property, we keep the changed value instead. More over for
`em.findOne()` with a PK condition, we try to detect whether it makes sense
to reload an entity, by comparing the options and already loaded property names.
In this step the `fields` and `populate` options are taken into account to
support both partial loading and lazy properties.

For complex conditions and `em.find()` we always do the query anyway, and we now
merge the newly loaded data from such queries in the same manner.